fun searchItemFromCategory(id) works

+++ app/src/main/java/ar/teamrocket/duelosmeli/model/Article.kt
@@ -6,7 +6,7 @@ data class Article (
     var id: String,
     var title: String,
     var pictures: List<ArticlePicture>,
-    var price: Long
+    var price: Double
 )
 
 data class ArticlePicture (
